+++ src/test/java/org/zalando/nakadi/controller/EventTypeControllerTest.java
@@ -731,8 +731,7 @@ public class EventTypeControllerTest extends EventTypeControllerTestCase {
         eventType.getSchema().setSchema("invalid-json");
 
         final Problem expectedProblem =
-                createInvalidEventTypeExceptionProblem("schema must be a valid json: Unexpected token 'invalid' " +
-                        "on line 1, char 1");
+                createInvalidEventTypeExceptionProblem("schema must be a valid json: Unexpected symbol 'i' at pos 1");
 
         postETAndExpect422WithProblem(eventType, expectedProblem);
     }